Add possibility to single step one iteration of game update
Add possibility to single step one iteration of game update. This
is a useful debug feature when combind with debug pause. For example
in the play window we can now 1. hit debug pause, 2. step the game
forward in small increments to see what happens in our game code.

+++ engine/test/main.cpp
@@ -1001,7 +1001,7 @@ public:
         mDevice->CleanGarbage(120, gfx::Device::GCFlags::Textures);
     }
 
-    virtual void Update(double dt) override
+    virtual void Update(float dt) override
     {
         const auto time_step = 1.0/60.0;
         const auto tick_step = 1.0/1.0;
